+++ src/queries/keyStore.ts
@@ -1,3 +1,4 @@
+import { LocationQueryParams } from '@/apis/myPlant/getMyAllPlant';
 import { createQueryKeyStore } from '@lukemorales/query-key-factory';
 
 export const keyStore = createQueryKeyStore({
@@ -9,5 +10,8 @@ export const keyStore = createQueryKeyStore({
     getAllLocation: null,
     updateLocation: null,
     deleteLocation: null,
+  },
+  myPlant: {
+    getMyAllPlant: (query?: LocationQueryParams) => ['getMyAllPlant', query],
   },
 });
