+++ src/types/MyPlant.ts
@@ -1,0 +1,9 @@
+interface Plant {
+  myPlantId: number;
+  nickname: string;
+  scientificName: string;
+  image: string;
+  waterRemainDay: number;
+  fertilizerRemainDay: number;
+}
+export default Plant;
